feat: discord notification for some signals
- Added a discord notification wrapper for signals functions
- Some refacto of signals, commands and tasks to decouple some stuff
- Changed Contest.event relation to OneToOne.
Nothing is planned for now to have more than one contest per event.
If the need arises, I will migrate from OneToOne to ForeignKey.
In the meantime, it's easier to work with a OneToOne, especially since many logic depend on the fact that there is only one contest per event.